chore(index): move main index under src folder
